feature: major cleanup and reorganization of pretraining module to improve readability
